[HWKMETRICS-756] Previous commits did not backport all changes for expired metrics job
